requirements.txt
Django~=4.1.3
django-crum==0.7.9